requirements.txt
plotly
numpy
pandas
matplotlib
setuptools
scipy
jupyterlab
dash
openpyxl
pyyaml